Key View Controller changes
Made key changes to each of the view controllers to allow data and references of view controllers to be passsed to other view controllers.

This sorted out many of the UX problems i was facing. such as the detail viewControllers  from the tablevView items of 'NavigationPanelVC' not being embedded in a navigtion controller, and the animation not being activated once a tableview item has been selected